wip ( Override colors for dark background)
