+++ app/build.gradle
@@ -63,6 +63,7 @@ dependencies {
     implementation 'androidx.compose.material3:material3:1.0.0-alpha02'
     implementation files('network')
     implementation project(path: ':network')
+    implementation 'com.google.android.material:material:1.8.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
